Feat: 스테이지 1 완성

+++ UNIJAMMAIN/Assets/Scripts/Obstacle/Phase/MonsterData.cs
@@ -4,7 +4,7 @@ public struct MonsterData
     public bool isIn;
     public Define.MonsterType monsterType;
     public float spawnDuration;
-    public int numInRow;
+    public float numInRow;
     public float speedUpRate;
 
     public float MovingToHolderTime =>spawnDuration*numInRow;
